pyproject.toml
[project]
name = "deepthink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "boto3>=1.42.65",
    "openai>=2.26.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.1.0",
    "pydantic-settings>=2.8.0",
]
